+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `force_cache` extension to enforce the request to be cached, ignoring the HTTP headers. (#117)
 - Fix issue where sqlite storage cache get deleted immediately. (#119)
+- Support float numbers for storage ttl. (#107)
 
 ## 0.0.18 (23/11/2023)